move binding redirect with assembly resolve to separate fixture
