+++ CefSharp.Test/CefSharp.Test.csproj
@@ -101,6 +101,8 @@
     <Compile Include="Framework\AsyncExtensionFacts.cs" />
     <Compile Include="Framework\ConcurrentMethodRunnerQueueFacts.cs" />
     <Compile Include="OffScreen\OffScreenBrowserBasicFacts.cs" />
+    <Compile Include="BindingRedirectFixture.cs" />
+    <Compile Include="BindingRedirectFixtureCollection.cs" />
     <Compile Include="CefSharpFixture.cs" />
     <Compile Include="CefSharpFixtureCollection.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
