+++ CefSharp.Test/CefSharpFixture.cs
@@ -4,7 +4,6 @@
 
 using System;
 using System.IO;
-using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 using CefSharp.OffScreen;
@@ -18,11 +17,6 @@ namespace CefSharp.Test
 
         public CefSharpFixture()
         {
-            // CefSharp requires a default AppDomain which means that xunit is not able
-            // to provide the correct binding redirects defined in the app.config
-            // so we have to provide them manually via AssemblyResolve
-            AppDomain.CurrentDomain.AssemblyResolve += CurrentDomainAssemblyResolve;
-
             SynchronizationContext.SetSynchronizationContext(new SynchronizationContext());
 
             scheduler = TaskScheduler.FromCurrentSynchronizationContext();
@@ -60,19 +54,6 @@ namespace CefSharp.Test
                     }
                 });
             }
-
-            AppDomain.CurrentDomain.AssemblyResolve -= CurrentDomainAssemblyResolve;
-        }
-
-        private Assembly CurrentDomainAssemblyResolve(object sender, ResolveEventArgs args)
-        {
-            var asemblyName = new AssemblyName(args.Name);
-            var path = Path.Combine(Environment.CurrentDirectory, asemblyName.Name + ".dll");
-            if (File.Exists(path))
-            {
-                return Assembly.LoadFrom(path);
-            }
-            return null;
         }
     }
 }
